backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
celery==5.4.0
redis==5.0.7
boto3==1.34.144
python-dotenv==1.0.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
marshmallow==3.21.3
ffmpeg-python==0.2.0
pillow==10.4.0
python-magic==0.4.27
sqlalchemy==2.0.31
alembic==1.13.2
pytest==8.3.2
pytest-flask==1.3.0
prometheus-flask-exporter==0.23.1
flower==2.0.1
